test disc encryption
