+++ .github/workflows/posture.yaml
@@ -26,7 +26,40 @@ jobs:
           apt-get update
           apt-get install -y build-essential libgtk-3-dev libwebkit2gtk-4.1-dev libayatana-appindicator3-dev librsvg2-dev patchelf libssl-dev libxdo-dev unzip protobuf-compiler libprotobuf-dev rpm
 
-      - name: Run test
+      - name: Test kernel version and unencrypted disk
         working-directory: src-tauri/
         run: |
-          cargo test --locked --lib enterprise::inspector::tests::linux::test_linux_kernel_version -- --ignored --exact
+          cargo test --locked --lib enterprise::inspector::tests::ci::linux::test_linux_kernel_version -- --ignored --exact
+          cargo test --locked --lib enterprise::inspector::tests::ci::linux::test_disk_encryption_status_unencrypted -- --ignored --exact
+
+      - name: Install LUKS tools
+        run: |
+          if command -v dnf >/dev/null; then dnf install -y cryptsetup util-linux;
+          elif command -v yum >/dev/null; then yum install -y cryptsetup util-linux;
+          else apt-get update && apt-get install -y cryptsetup util-linux;
+          f
+
+      - name: Setup an encrypted partition
+        run: |
+          set -euo pipefail
+          img="$(mktemp /tmp/defguard-luks.XXXXXX)"
+          key="$(mktemp /tmp/defguard-luks-key.XXXXXX)"
+          loopdev=""
+          cleanup() {
+            if [ -n "${mapper_opened:-}" ]; then cryptsetup close defguard-posture-test || true; fi
+            if [ -n "$loopdev" ]; then losetup -d "$loopdev" || true; fi
+            rm -f "$img" "$key"
+          }
+          trap cleanup EXIT
+          truncate -s 64M "$img"
+          printf 'defguard-ci-test-key' > "$key"
+          loopdev="$(losetup --find --show "$img")"
+          cryptsetup luksFormat --batch-mode --type luks2 "$loopdev" "$key"
+          cryptsetup open "$loopdev" defguard-posture-test --key-file "$key"
+          mapper_opened=1
+          lsblk -Jo NAME,FSTYPE,TYPE
+
+      - name: Test encrypted disk
+        working-directory: src-tauri/
+        run: |
+          cargo test --locked --lib enterprise::inspector::tests::ci::linux::test_disk_encryption_status_encrypted -- --ignored --exact
