+++ src-tauri/src/enterprise/inspector/tests/ci/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(target_os = "linux")]
+mod linux;
