+++ src-tauri/src/enterprise/inspector/tests/mod.rs
@@ -1,3 +1,4 @@
+mod ci;
 #[cfg(target_os = "linux")]
 mod linux;
 #[cfg(target_os = "macos")]
